use sd-bus for IPC with NetworkManager instead of nmcli
This avoids the overhead of spawning a process, a process which:
- uses a less efficient D-Bus library
- gathers more information than is needed by vanilla

This should hopefully speed things up the tiniest bit. It also no longer relies
on the presence of nmcli in path to tell if NetworkManager is present/running.

Note that I am using __attribute__((__cleanup__(x))), which is a GNU extension
(also supported by clang and ICC), as this saves me having to remember to free
everything, and should help prevent memory leaks. I hope this is not an issue.