disconnect and reactivate Wi-Fi connection on startup and shutdown
iwd, when managed by NetworkManager, does not automatically disconnect from the
network when we tell NetworkManager to unmanage the network, unlike
wpa_supplicant. A workaround for this is to just tell NetworkManager to
disconnect from the network, and it will tell iwd to do so.

We also store the network configuration until (our) shutdown so it can be
reapplied, as NetworkManager doesn't seem to autoconnect to the same network
after you disconnect, even if it has stopped and started managing the network.